requirements.txt
absl-py==2.4.0
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
audioread==3.1.0
braceexpand==0.1.7
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
cuda-bindings==13.1.1
cuda-pathfinder==1.3.4
cycler==0.12.1
cytoolz==1.1.0
datasets==4.5.0
decorator==5.2.1
dill==0.4.0
editdistance==0.8.1
einops==0.8.2
executing==2.2.1
ffmpeg==1.4
fiddle==0.3.0
filelock==3.24.3
fonttools==4.61.1
frozenlist==1.8.0
gitdb==4.0.12
gitpython==3.1.46
graphviz==0.21
greenlet==3.3.1
grpcio==1.78.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.4.1
hydra-core==1.3.2
idna==3.11
indic-numtowords==1.1.0
inflect==7.5.0
intervaltree==3.2.1
ipython==9.10.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jiwer==3.1.0
joblib==1.5.3
kaldi-python-io==1.2.2
kaldialign==0.9.1
kiwisolver==1.4.9
lazy-loader==0.4
levenshtein==0.27.3
lhotse==1.31.1
libcst==1.8.6
librosa==0.11.0
lilcom==1.8.2
llvmlite==0.46.0
mako==1.3.10
markdown==3.10.2
markdown-it-py==4.0.0
markupsafe==2.1.5
marshmallow==4.2.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mediapy==1.1.6
ml-dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.18
networkx==3.6.1
numba==0.63.1
numba-cuda==0.15.1
numexpr==2.13.1
numpy==2.3.0
nv-one-logger-core==2.3.1
nv-one-logger-pytorch-lightning-integration==2.3.1
nv-one-logger-training-telemetry==2.3.1
omegaconf==2.3.0
onnx==1.20.1
optuna==4.7.0
overrides==7.7.0
packaging==26.0
pandas==3.0.0
parso==0.8.6
peft==0.18.1
pillow==12.0.0
platformdirs==4.5.1
pooch==1.9.0
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==5.29.6
psutil==7.2.2
pure-eval==0.2.3
pyannote-core==6.0.1
pyannote-database==6.1.1
pyannote-metrics==4.0.0
pyarrow==23.0.0
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydub==0.25.1
pygments==2.19.2
pyloudnorm==0.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-levenshtein==0.27.3
pytorch-lightning==2.6.1
pyyaml==6.0.3
rapidfuzz==3.14.3
regex==2026.1.15
requests==2.32.5
resampy==0.4.3
rich==14.3.2
ruamel-yaml==0.19.1
sacremoses==0.1.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentencepiece==0.2.1
sentry-sdk==2.52.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.5.0
soxr==1.0.0
sqlalchemy==2.0.46
stack-data==0.6.3
strenum==0.4.15
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
toolz==1.1.0
tqdm==4.67.3
traitlets==5.14.3
transformers==5.2.0
typeguard==4.4.4
typer==0.24.0
typer-slim==0.24.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
wandb==0.24.2
wcwidth==0.6.0
webdataset==1.0.2
werkzeug==3.1.5
wget==3.2
whisper-normalizer==0.1.12
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
